requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
prisma==0.15.0
httpx==0.27.2
python-multipart==0.0.9
librosa==0.10.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.3.2
soundfile==0.12.1
